+++ mathics/session.py
@@ -12,6 +12,7 @@ In particular we provide:
 import os.path as osp
 from typing import Optional
 
+from mathics.builtin.system_init import initialize_system
 from mathics.core.definitions import autoload_files
 from mathics.core.parser import parse, MathicsSingleLineFeeder
 from mathics.core.definitions import Definitions
@@ -63,6 +64,7 @@ class MathicsSession:
         form="InputForm",
         character_encoding=Optional[str],
     ):
+        initialize_system()
         if character_encoding is not None:
             mathics.settings.SYSTEM_CHARACTER_ENCODING = character_encoding
         self.form = form
